Refactor parser for class and separate validation
